+++ NEWS.md
@@ -17,6 +17,7 @@
 - Improved reference documentation for `analysis-information` (issue #33)
 - Handbook: detailed overview of all workflow functions and classes (issue #41, special thanks to @@hechth)
 - Fixed: annotations slot for `featureGroups` was not updated when removing groups with `delete()` (except sets workflows)
+- Fixed: better handle missing spectrum data with spectrumSimilarity()
 
 
 # patRoon 2.0